pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bsim"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Demi", email = "bjaiye1@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = []

[project.urls]
Documentation = "https://github.com/Demi/bsim#readme"
Issues = "https://github.com/Demi/bsim/issues"
Source = "https://github.com/Demi/bsim"

[tool.hatch.version]
path = "src/bsim/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/bsim tests}"

[tool.coverage.run]
source_pkgs = ["bsim", "tests"]
branch = true
parallel = true
omit = [
  "src/bsim/__about__.py",
]

[tool.coverage.paths]
bsim = ["src/bsim", "*/bsim/src/bsim"]
tests = ["tests", "*/bsim/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
testpaths = [
  "tests",
]
python_files = [
  "test_*.py",
]
addopts = "-q"

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "pytest-cov>=4",
  "pyyaml>=6",
  "tomli>=2; python_version < '3.11'",
]
ui = [
  "fastapi>=0.110,<1",
  "uvicorn>=0.23",
]
# Everything
all = [
  "bsim[dev,ui]",
]

[tool.hatch.build]
include = [
  "src/bsim/simui/static/**",
]
exclude = [
  "src/bsim/simui/_frontend/**",
]
